backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
langchain==0.0.347
openai==1.3.0
youtube-transcript-api==0.6.1
pinecone-client==3.0.1
python-dotenv==1.0.0
chromadb==0.4.22
tiktoken==0.5.2
